Move prf_expand{_vec} into a submodule for clarity
